Refactor configuration and update flake dependencies
The commit updated the flake dependencies, refactored the configuration format and fixed issues with the rebuild script. This includes further organizing directories in the home configuration, modifying the 'home-manager' and 'xserver' settings, and correcting the syntax error in 'home.nix' and 'configuration.nix'. The configurations for the 'flake.nix' and 'rebuild.sh' files were also adjusted.