Remove EWW config and add new scripts and configurations
This commit removes the EWW configuration in the home.nix. It also includes several new scripts and configurations which are the addition of .editorconfig, updgrade.sh, prettier configuration in .prettierc and package.json, .gitignore file, and changes in several other configurations as well as scripts. EWW's removal is due to the incorporation of a new bar configuration, makeBar, in the config.js.